v6: Using a hidden parent node => 80K balls
In version 5, we apply 'display:none' to each ball to hide it -> restyle + reflow + repaint.
In version 6, we move a ball to a hidden node to hide it -> reflow + repaint. As a result, we can animate 80,000 balls smoothly.
It took a lot of time to generate balls at the beginning, so we add another method using 'cloneNode' function.
Note: 'movers = container.children' won't work anymore. Because 'movers' always refers to 'container.children', then when we move some balls to 'hidden', number of balls in movers will be reduced. An alternative way is 'movers = container.querySelectorAll(".mover")', which is slow.
Note: 'camera' debug is not necessary anymore, so I won't update its position in 'update_v*'.