v7: Pre-calculate visible index => 140K balls
In version 6, we check every ball to decide it should be visible or hidden.
In version 7, we find the 1st and last visible balls first to define a visible range. Then we hide all other balls not in visible range without checking. By doing that, we can animate 140K balls smoothly.
Note: Be careful with 'from' & 'to' parameters because we have PADDING => visible range might be out of 'movers' array.
P/S: In line 270, 282, 300 I wrote 'if' statements in 2 line without brackets {}. It's not my usual coding convention, but it will be useful in next versions.