v7b: Separate move & show => 140K balls
In version 7 (or 7a), we move a ball then hide it immediately before processing next ball in 'batchDraw' function.
In version 7b, we move all the needed balls first, then show them all at once later to see if there's any difference but almost got the same result as 7a.
So we need to find another way to improve version 7.