v8b: Generate hidden balls by default
In version 8 (or 8a), we need 13s to generate 1M balls, and 12s more to paint the 1st frame => 25s of waiting.
We know what the 1st frame looks like.
So in version 8b, we attach visible balls to container and hidden balls to hidden.
By doing that, 1st call to 'update_v8' will result in 'identical' case => 0s for re-painting => 13s of waiting (50%)