v8c: Create visible balls first, hidden ones later
In version 8a & 8b, it took 13s to paint the 1st frame, it's too long.
Most of the time it spent is to create hidden balls.
So in version 8c, we just create visible balls and display the 1st frame immediately.
We delay creating hidden balls until next frame.
But the time we need to create hidden balls is still long, so we need to disable "Play / Pause" button first then enable it when initialization is done.
More than that, we need to disable the 1st call to rAF until all the balls are ready.