v8.1b: Use cssText => 200 balls @ 15FPS
In version 8.1 (or 8.1a), we animated 'left', 'width', & 'height' separately by 3 commands.
So in version 8.1b, we combine all 3 commands into 1 by using 'cssText'.
Because 'cssText' will overwrite all inline style, so we need to rewrite 'top' attribute too.
The performance is still terrible.
Tab 'Performance' shows that GPU takes a lot time to run (but I have no idea how to fix this).
Tab 'Performance monitor' shows that there are a lots of Reflow (Layouts / sec), because we update 4 attributes which are able to reflow.