v8.3: Add will-change => 1K balls
In theory, opacity (& transform) attribute can be processed by GPU in 'composite' phase very efficiently.
But it only happens when element is on its own layer.
Otherwise, other change in the same layer will force GPU to re-render the whole layer again.
All our balls are in the same layer. So when each ball changes opacity, the "balls" layer have to be re-composited.
That's why GPU took so long to process in v8.2b.
By adding 'will-change' attribute to each ball, we tells GPU that each ball should be in a separate layer.
So any changes from a ball won't affect other balls.
With this tweak, we can animate 1,000 balls (or more) smoothly.