v9.2: Add delay to each ball
Just by delaying the start of animation, we can mimic JS animation in v8.4.
But if the 2nd ball is delayed 120ms, then the 101th ball will be delayed 12s.
It is too long for user to wait for the moment all the balls moving.
So we need to calculate the overlapped time needed to delay each ball group by finding actual time needed for a ball to complete a loop (6.28s).
In the screenshot 'v9.2-1K balls', we can see that GPU do all the work to animate 1,000 balls at 60FPS smoothly