v9.3: Make button "Play/Pause" work
In previous versions, there's no way to pause and resume playing animations.
In this version, we add this feature by using attribute 'animation-play-state'.
But applying this attribute to all the balls one by one is slow.
So we need to think of a way that can be applied immediately for all the ball at the same time.
The solution is 'CSS selector', we can declare that all the balls will stop animating if their parent has class 'pause'.