v9.5: Dynamic position in animation
In previous version 9.x, all attributes are fixed.
Opacity and scale can be fixed.
But 'translateX' should be changed with the window size (window.innerHeight).
So in this version, I use CSS custom variable to change this attribute dynamically.
Performance doesn't change with 9.4, but the overall is not as smooth as v8.4.